Make base class for RepositoryHelper tests
